# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restaurant Management System API",
        "description": "API documentation for Restaurant Management System Backend",
        "contact": {
            "email": "admin@restaurant.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Restaurant Management API Server"
        }
    ],
    "paths": {
...
        "/api/auth/order-items/status": {
            "put": {
                "tags": [
                    "OrderItems"
                ],
                "summary": "Cập nhật trạng thái nhiều OrderItem",
                "operationId": "f3c123c3213e9525cf6df093016760b0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "items": {
                                        "description": "Key-value: orderItemId => status",
                                        "type": "object",
                                        "additionalProperties": {
                                            "description": "0=Ordered, 1=Cooking, 2=Served, 3=Cancelled",
                                            "type": "integer",
                                            "enum": [
                                                0,
                                                1,
                                                2,
                                                3
                                            ],
                                            "property": null
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order items and order status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "OrderItem ID",
                                                "type": "string"
                                            },
                                            "order_id": {
                                                "description": "Order ID",
                                                "type": "string"
                                            },
                                            "status": {
                                                "description": "OrderItem status",
                                                "type": "integer",
                                                "enum": [
                                                    0,
                                                    1,
                                                    2,
                                                    3
                                                ]
                                            },
                                            "quantity": {
                                                "type": "integer"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "total_price": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "notes": {
                                                "type": "string",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validate lỗi input"
                    },
                    "403": {
                        "description": "Không có quyền"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/order-items/add-order": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Thêm order mới cùng danh sách món",
                "operationId": "de5ae79acb6085ba889e1d0039a8294e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id",
                                    "items"
                                ],
                                "properties": {
                                    "order_id": {
                                        "description": "ID của order",
                                        "type": "string"
                                    },
                                    "items": {
                                        "description": "Danh sách món ăn trong order",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "dish_id",
                                                "name_dish",
                                                "price",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "dish_id": {
                                                    "description": "ID món ăn",
                                                    "type": "string"
                                                },
                                                "name_dish": {
                                                    "description": "Tên món ăn",
                                                    "type": "string"
                                                },
                                                "price": {
                                                    "description": "Đơn giá món",
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "quantity": {
                                                    "description": "Số lượng",
                                                    "type": "integer"
                                                },
                                                "status": {
                                                    "description": "Trạng thái món (0=Open,1=In-progress,2=Served,3=Cancelled)",
                                                    "type": "integer",
                                                    "enum": [
                                                        0,
                                                        1,
                                                        2,
                                                        3
                                                    ],
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order và các món đã được tạo thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "order": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string"
                                                        },
                                                        "status": {
                                                            "type": "integer"
                                                        },
                                                        "total_amount": {
                                                            "type": "number",
                                                            "format": "float"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "description": "OrderItem ID",
                                                                "type": "string"
                                                            },
                                                            "dish_id": {
                                                                "type": "string"
                                                            },
                                                            "name_dish": {
                                                                "type": "string"
                                                            },
                                                            "price": {
                                                                "type": "number",
                                                                "format": "float"
                                                            },
                                                            "quantity": {
                                                                "type": "integer"
                                                            },
                                                            "total_price": {
                                                                "type": "number",
                                                                "format": "float"
                                                            },
                                                            "status": {
                                                                "type": "integer",
                                                                "enum": [
                                                                    0,
                                                                    1,
                                                                    2,
                                                                    3
                                                                ]
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Order và các món đã được tạo thành công."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validate lỗi input"
                    },
                    "403": {
                        "description": "Không có quyền"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}